+++ src/TrackWrapper.js
@@ -3,6 +3,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { InfoContext, ACTION } from "./utils/contexts.js";
 import TrackRowInfo from './TrackRowInfo.js';
 import TrackRowHighlight from './TrackRowHighlight.js';
+import TrackRowZoomOverlay from './TrackRowZoomOverlay.js';
 
 // TODO: remove the below fakedata import.
 //       see https://github.com/hms-dbmi/cistrome-higlass-wrapper/issues/26
@@ -17,12 +18,14 @@ import fakedata from './demo/fakedata/index.js';
  * @prop {function} onAddTrack The function to call upon a track insertion.
  * @prop {function} onSortRows The function to call upon a sort interaction.
  * @prop {function} onSearchRows The function to call upon a search interaction.
+ * @prop {function} onZoomRows The function to call upon a vertical zoom interaction.
  * @prop {function} onFilterRows The function to call upon a filer interaction.
  * @prop {function} onMetadataLoad The function to call upon rowInfo is set to Context.
+ * @prop {boolean} isWheelListening Whether or not to listen for wheel events for vertical zooming.
  * @prop {function} drawRegister The function for child components to call to register their draw functions.
  */
 export default function TrackWrapper(props) {
-    const { 
+    const {
         options, 
         multivecTrack,
         multivecTrackViewId,
@@ -30,8 +33,10 @@ export default function TrackWrapper(props) {
         onAddTrack,
         onSortRows,
         onSearchRows,
+        onZoomRows,
         onFilterRows,
         onMetadataLoad,
+        isWheelListening,
         drawRegister
     } = props;
 
@@ -158,6 +163,14 @@ export default function TrackWrapper(props) {
                 selectedRows={selectedRows}
                 highlitRows={highlitRows}
                 drawRegister={drawRegister}
+            />
+            <TrackRowZoomOverlay
+                trackX={trackX}
+                trackY={trackY}
+                trackHeight={trackHeight}
+                trackWidth={trackWidth}
+                isWheelListening={isWheelListening}
+                onZoomRows={onZoomRows}
             />
         </div>
     );